requirements.txt
Flask
google-generativeai
gtts
pygame
